# src/api/docs/swaggerbackup.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "DIN DIN API",
    "description": "DIN DIN apis"
  },
  "basePath": "/",
  "schemes": [
    "https",
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
     "/practice/profile": {
      "post": {
        "tags": [
          "Practice"
        ],
        "summary": "Add profile",
        "description": "Add user profile",
        "operationId": "profile",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/HeaderVersion"
          },
          {
            "$ref": "#/parameters/HeaderLanguage"
          },
          {
            "$ref": "#/parameters/HeaderPlatform"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/user_profile"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User profile in system",
            "schema": {
              "$ref": "#/definitions/responsesOk"
            }
          },
          "500": {
            "description": "Unexpected error.",
            "schema": {
              "$ref": "#/definitions/responsesError"
            }
          }
        }
      }
    },
...
    "/practice/password": {
      "put": {
        "tags": [
          "Practice"
        ],
        "summary": "Change password",
        "description": "Change passord",
        "operationId": "password",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/HeaderVersion"
          },
          {
            "$ref": "#/parameters/HeaderLanguage"
          },
          {
            "$ref": "#/parameters/HeaderPlatform"
          },
          {
            "$ref": "#/parameters/HeaderAuth"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/change_password"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Change password in system",
            "schema": {
              "$ref": "#/definitions/responsesOk"
            }
          },
          "500": {
            "description": "Unexpected error.",
            "schema": {
              "$ref": "#/definitions/responsesError"
            }
          }
        }
      }
    },
...
    "/concierges/profile": {
      "get": {
        "tags": [
          "Concierges"
        ],
        "summary": "Get Concierge profile",
        "description": "Get Concierge profile",
        "operationId": "profile",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/HeaderVersion"
          },
          {
            "$ref": "#/parameters/HeaderLanguage"
          },
          {
            "$ref": "#/parameters/HeaderPlatform"
          },
          {
            "$ref": "#/parameters/HeaderAuth"
          }
        ],
        "responses": {
          "200": {
            "description": "Get concierge profile",
            "schema": {
              "$ref": "#/definitions/responsesOk"
            }
          },
          "500": {
            "description": "Unexpected error.",
            "schema": {
              "$ref": "#/definitions/responsesError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
     "user_profile": {
      "type": "object",
      "required": [
        "_id",
        "shippingAddress1",
        "shippingAddress2",
        "city",
        "zipCode",
        "state",
        "secondaryPhone",
        "licenseNumber"
        
      ],
      "properties": {
        "_id": {
          "type": "string",
          "example": "5ed660d63bbd282d947af1c6" 
        },
        "shippingAddress1": {
          "type": "string",
          "example": "Avenue street"
        },
        "shippingAddress2": {
          "type": "string",
          "example": "3 mall"
        },
        "state": {
          "type": "string",
          "example": "5e15b6f7dbc1a60870ed864e"
        },
        "city": {
          "type": "string",
          "example": "Florida"
        },
         "zipCode": {
          "type": "string",
          "example": "123456"
        },
        "secondaryPhone": {
          "type": "string",
         
          "example": "(123) 345 7890"
        },
       
        "licenseNumber": {
          "type": "String",
          "example": "123Fgj"
        }
      }
    },
...
     "change_password": {
      "type": "object",
      "required": [
         "currentPassword",
         "newPassword"
      ],
      "properties": {
        "currentPassword": {
          "type": "string",
          "description":"Current password" ,
          "format": "password",
          "example": "12345"
        },
         "newPassword": {
          "type": "string",
          "description":"Current password" ,
          "format": "password",
          "example": "123456"
        }
       
      }
    },
...
    "responsesOk": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "code": {
          "type": "string"
        },
        "msg": {
          "type": "string"
        },
        "data": {
          "type": "object"
        }
      }
    },
    "responsesError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "code": {
          "type": "string"
        },
        "msg": {
          "type": "string"
        },
        "error": {
          "type": "object"
        }
      }
    },
...
  },
  "parameters": {
    "HeaderVersion": {
      "name": "X-DinDin-Version",
      "in": "header",
      "description": "X-DinDin-Version",
      "required": true,
      "type": "string",
      "default": "1.0.0"
    },
    "HeaderLanguage": {
      "name": "Accept-Language",
      "in": "header",
      "description": "Accept-Language",
      "required": true,
      "type": "string",
      "default": "en"
    },
    "HeaderPlatform": {
      "name": "X-DinDin-Platform",
      "in": "header",
      "description": "android or ios or web",
      "required": true,
      "type": "string",
      "default": "android"
    },
    "HeaderAuth": {
      "name": "Authorization",
      "in": "header",
      "description": "Authorization",
      "required": true,
      "type": "string"
    },
...
  }
}